optimized dashborad page code
