+++ web/AASTU-web-group-1/aastu.web.g1.bank-dashboard/app/dashboard/setting/component/Security.tsx
@@ -1,8 +1,5 @@
 "use client";
 
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
 import { Button } from "@/components/ui/button";
 import {
   Form,
@@ -14,9 +11,12 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { Switch } from "@/components/ui/switch";
+import { zodResolver } from "@hookform/resolvers/zod";
 import { getSession } from "next-auth/react";
-import { toast } from "sonner";
 import { useState } from "react";
+import { useForm } from "react-hook-form";
+import { toast } from "sonner";
+import { z } from "zod";
 
 const Security = () => {
   const [loading, setLoading] = useState(false);
@@ -57,7 +57,7 @@ const Security = () => {
       setLoading(true);
 
       const res = await fetch(
-        "https://bank-dashboard-1tst.onrender.com/auth/change_password",
+        "https://bank-dashboard-o9tl.onrender.com/auth/change_password",
         {
           method: "POST",
           body: JSON.stringify(formData),
